+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+0.74 -> 0.75
+  - Major refracturing
+  - Improvement of Documentation
+  - New play method for quick signal playback
+  - Added resample method to signal
+
 0.73 -> 0.74
   - Improvements of the Documentation.
   - Implemented several spectrograms in the time_frequency sub-module of the Signal class.